bvms-confluence/requirements.txt
# Basic requirements
requests>=2.28.0
python-dotenv>=1.0.0
beautifulsoup4>=4.11.0

# Optional advanced features (uncomment to enable)
atlassian-python-api>=3.41.0
markdownify>=0.11.6
